% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{atlantis}
\alias{atlantis}
\title{Atlantis}
\format{
HEX color elements that match the RV Atlantis
...
}
\source{
\url{https://www.whoi.edu}
}
\usage{
atlantis
}
\description{
A list of HEX colors, n = 9
}
\keyword{datasets}
