% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bikepath}
\alias{bikepath}
\title{Bikepath}
\format{
HEX color elements that represent the bikepath
...
}
\source{
\url{https://www.whoi.edu}
}
\usage{
bikepath
}
\description{
A list of HEX colors, n = 7, that represent the essence of the Shining Sea Bikeway
}
\keyword{datasets}
