% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bog}
\alias{bog}
\title{Cranberry bog}
\format{
HEX color elements that represent a cranberry bog
...
}
\source{
\url{https://www.cranberries.org/how-cranberries-grow}
}
\usage{
bog
}
\description{
A list of HEX colors, n = 6, that match the majestic cranberry bogs of Massachusetts
}
\keyword{datasets}
