% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dock}
\alias{dock}
\title{Dock color palette}
\format{
List of colors that represent dock.
...
}
\usage{
dock
}
\description{
Image of a dock from Eel Pond in Woods Hole. n =9 colors
}
\keyword{datasets}
