% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{eelpond_winter}
\alias{eelpond_winter}
\title{Eel Pond in winter}
\format{
HEX color elements that represent the winter in Eel pond
...
}
\source{
\url{https://woodshole.com/webcams/eel-pond-from-mbl/}
}
\usage{
eelpond_winter
}
\description{
A list of HEX colors, n = 6, that convey the beauty of fresh snowfall on Eel Pond in Woods Hole.
}
\keyword{datasets}
