% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{knob1}
\alias{knob1}
\title{Color palette from the Knob at sunset}
\format{
List of colors that represent tee Knob
...
}
\usage{
knob1
}
\description{
Image of the Knob beach
}
\keyword{datasets}
