% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{long_pond}
\alias{long_pond}
\title{Long Pond color palette}
\format{
List of colors that represent Long Pond, HEX
...
}
\usage{
long_pond
}
\description{
Image of Long Pond in the fall
}
\keyword{datasets}
