% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{marsh}
\alias{marsh}
\title{Marsh color palette}
\format{
List of colors that represent marsh essence!
...
}
\usage{
marsh
}
\description{
Image of Sippewissett marsh n =9 colors
}
\keyword{datasets}
