% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{naushon}
\alias{naushon}
\title{Naushon Island, Hadley Harbor}
\format{
List of colors that represent Naushon Island
...
}
\usage{
naushon
}
\description{
Image from Naushon Island
}
\keyword{datasets}
