% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rocky_beach}
\alias{rocky_beach}
\title{Rocky beach}
\format{
HEX color elements for rocky beach
...
}
\usage{
rocky_beach
}
\description{
You remember the summertime in Woods Hole. Going to beach may mean a visit to a rocky beach. A list of HEX colors, n = 6, to bring you to the beach.
}
\keyword{datasets}
