% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_color.R
\name{show_color}
\alias{show_color}
\title{Prints grid of colors}
\usage{
show_color(color)
}
\arguments{
\item{color}{input color name}
}
\value{
Grid using 'scales' with hex code and color
}
\description{
Prints grid of colors
}
\examples{
show_color(atlantis)
}
