% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{siders_filters}
\alias{siders_filters}
\title{Siders filters color palette}
\format{
List of colors that represent Siders, HEX
...
}
\usage{
siders_filters
}
\description{
Filters collected from depth profile at Siders pond
}
\keyword{datasets}
