% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{siders_summer}
\alias{siders_summer}
\title{Siders color palette in summer}
\format{
List of colors that represent Siders, HEX
...
}
\usage{
siders_summer
}
\description{
Image of Siders pond in the summer, doing field work.
}
\keyword{datasets}
