% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sunset_winter}
\alias{sunset_winter}
\title{Sunset winter}
\format{
HEX color elements for wintertime sunset
...
}
\source{
\url{}
}
\usage{
sunset_winter
}
\description{
Sunsets on the cape are not to be missed. Even in the winter. 6 HEX color codes to represent that winter sunset vibe.
}
\keyword{datasets}
