% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{tulips}
\alias{tulips}
\title{Tulips}
\format{
HEX color elements for tulips
...
}
\usage{
tulips
}
\description{
A sign of warmer weather. 9 HEX color code list that will bring you there - there, being the tulips that pop up in late spring.
}
\keyword{datasets}
