% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wefa_sun}
\alias{wefa_sun}
\title{West Falmouth Sunset}
\format{
HEX color elements for West Falmouth sunset
...
}
\usage{
wefa_sun
}
\description{
You've raced to Chappys to catch the sunset, right? 8 HEX color codes to bring the West Falmouth sunset to you.
}
\keyword{datasets}
