% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{whoi}
\alias{whoi}
\title{WHOI logo}
\format{
HEX color elements that match the logo
...
}
\source{
\url{https://www.whoi.edu}
}
\usage{
whoi
}
\description{
A list of HEX colors, n = 4
}
\keyword{datasets}
