% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{whoi2}
\alias{whoi2}
\title{WHOI logo 2}
\format{
HEX color elements that match the logo
...
}
\source{
\url{https://www.whoi.edu}
}
\usage{
whoi2
}
\description{
A list of HEX colors, n = 6
}
\keyword{datasets}
