% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{whoisec}
\alias{whoisec}
\title{WHOI secondary}
\format{
HEX color elements that are secondary.
...
}
\source{
\url{https://www.whoi.edu}
}
\usage{
whoisec
}
\description{
A list of HEX colors, n = 4, that are the WHOI secondary color palette
}
\keyword{datasets}
