% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-wh.R, R/data.r
\docType{data}
\name{woodshole}
\alias{woodshole}
\title{Complete list of woods hole color palettes}
\format{
An object of class \code{list} of length 11.

List of all Woods Hole color palettes available in the package.
...
}
\source{
\url{}
}
\usage{
woodshole

woodshole
}
\description{
Complete list of Woods Hole color palettes available in package.
}
\keyword{datasets}
