% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{woodshole}
\alias{woodshole}
\title{Woods Hole color palette}
\format{
List of all Woods Hole color palettes available in the package.
...
}
\source{
\url{}
}
\usage{
woodshole
}
\description{
Complete list of Woods Hole color palettes available in package.
}
\keyword{datasets}
